quick update for rendering py
